feat: Branching input component
